requirements.txt
streamlit==0.59.0
numpy==1.18.1
opencv-contrib-python-headless==4.2.0.34
Pillow==7.1.2
requests==2.22.0
Flask==1.1.1
Flask_Cors==3.0.8
gunicorn==19.9.0